@@ -1,7 +1,7 @@
 package com.helesto.core;
 
-import javax.inject.Inject;
-import javax.inject.Singleton;
+import jakarta.inject.Inject;
+import jakarta.inject.Singleton;
 
 import com.helesto.service.ExecutionReportService;
 
