@@ -2,10 +2,10 @@ package com.helesto.dao;
 
 import java.util.List;
 
-import javax.enterprise.context.ApplicationScoped;
-import javax.inject.Inject;
-import javax.persistence.EntityManager;
-import javax.persistence.TypedQuery;
+import jakarta.enterprise.context.ApplicationScoped;
+import jakarta.inject.Inject;
+import jakarta.persistence.EntityManager;
+import jakarta.persistence.TypedQuery;
 
 import com.helesto.model.MessageLogIncomingEntity;
 import com.helesto.model.MessageLogOutgoingEntity;
