@@ -3,8 +3,8 @@ package com.helesto.service;
 import java.util.ArrayList;
 import java.util.List;
 
-import javax.enterprise.context.RequestScoped;
-import javax.inject.Inject;
+import jakarta.enterprise.context.RequestScoped;
+import jakarta.inject.Inject;
 
 import com.helesto.core.Trader;
 import com.helesto.dao.MessageLogDao;
