@@ -1,8 +1,8 @@
 package com.helesto;
 
-import io.quarkus.test.junit.NativeImageTest;
+import io.quarkus.test.junit.QuarkusIntegrationTest;
 
-@NativeImageTest
+@QuarkusIntegrationTest
 public class SettingsRestTestIT extends SessionRestTest {
 
     // Execute the same tests but in native mode.
